refactor: move app-component to shared directory
