+++ packages/spectacular/src/lib/application-testing/application-harness/create-application-harness.spec.ts
@@ -1,6 +1,6 @@
 import { APP_BOOTSTRAP_LISTENER, APP_INITIALIZER, FactoryProvider, Injectable, NgModule } from '@angular/core';
 
-import { SpectacularAppComponent } from '../app-component/spectacular-app.component';
+import { SpectacularAppComponent } from '../../shared/app-component/spectacular-app.component';
 import { createApplicationHarness } from './create-application-harness';
 
 let bootstrapped = false;
