+++ packages/spectacular/src/lib/application-testing/application-harness/create-application-harness.ts
@@ -2,8 +2,8 @@ import { NgZone } from '@angular/core';
 import { ComponentFixtureAutoDetect, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { SpectacularAppComponent, spectacularAppTag } from '../app-component/spectacular-app.component';
-import { SpectacularAppModule } from '../app-component/spectacular-app.module';
+import { SpectacularAppComponent, spectacularAppTag } from '../../shared/app-component/spectacular-app.component';
+import { SpectacularAppModule } from '../../shared/app-component/spectacular-app.module';
 import { bootstrapComponent } from '../util-bootstrapping/bootstrap-component';
 import { CreateApplicationHarnessOptions } from './create-application-harness-options';
 import { SpectacularApplicationHarness } from './spectacular-application-harness';
