+++ packages/spectacular/src/lib/feature-testing/feature-harness/create-feature-harness.spec.ts
@@ -4,7 +4,7 @@ import { Component, Injectable, NgModule } from '@angular/core';
 import { ExtraOptions, ROUTER_CONFIGURATION, RouterModule } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { SpectacularAppComponent } from '../../application-testing/app-component/spectacular-app.component';
+import { SpectacularAppComponent } from '../../shared/app-component/spectacular-app.component';
 import { featurePathToken } from '../configuration/feature-path.token';
 import { createFeatureHarness } from './create-feature-harness';
