+++ packages/spectacular/src/lib/shared/app-component/spectacular-app.component.spec.ts
