+++ packages/spectacular/src/lib/shared/app-component/spectacular-app.module.ts
